ooops.   Fixed Frog Lives display glitch.
On entry to the game screen it reports 0 Lives instead of 3.  Internally it is correctly starting with 3.   The counter was not updated on the screen until the first frog death.   Fixed this to call the Lives and Saved frogs update to the screen at the transition to the game screen.