Refining splash screens
Fixed accidental DLI addresses  exceeding a page.
Title screens seems good.
Game screen seems good.
Or so the emulators suggest.
Working on the splash screen refinements.... The splash screen color animations are higher on screen than the PM text labels on the Title and Game screens, so there were still bits of these appearing on the splash screen.  This required a special DLI set early on the Splash screens to force HPOS to zero for all object to get rid of them.  Later on the screen the P/M objects can be set to the specs from table entry+2 if there is an animated object on the screen. 
Still working on fixing the tombstone on the Game over screen.   Also need to fix things  to put the Frog back on the Win screen.